requirements.txt
aiosqlite    ~=0.17.0
ircrobots    ~=0.6.1
pyyaml       ~=5.4.1

